[MIG] stock_picking_back2draft: Migration to 13.0
