+++ stock_picking_back2draft/__manifest__.py
@@ -5,7 +5,7 @@
     "summary": "Reopen cancelled pickings",
     "version": "13.0.1.0.0",
     "category": "Warehouse Management",
-    "website": "https://www.agilebg.com",
+    "website": "https://github.com/OCA/stock-logistics-workflow",
     "author": "Agile Business Group, Odoo Community Association (OCA)",
     "license": "AGPL-3",
     "application": False,
