+++ stock_picking_back2draft/models/__init__.py
@@ -1,4 +1,5 @@
 # © 2016 Lorenzo Battistini - Agile Business Group
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from . import stock
+from . import stock_move
+from . import stock_picking
