+++ stock_picking_back2draft/models/stock_picking.py
@@ -1,0 +1,12 @@
+# © 2016 Lorenzo Battistini - Agile Business Group
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from odoo import models
+
+
+class StockPicking(models.Model):
+    _inherit = "stock.picking"
+
+    def action_back_to_draft(self):
+        moves = self.mapped("move_lines")
+        moves.action_back_to_draft()
