+++ stock_picking_back2draft/readme/CONTRIBUTORS.rst
@@ -1,0 +1,3 @@
+* Lorenzo Battistini <lorenzo.battistini@agilebg.com>
+* Iván Montagud <ivan@studio73.es>
+* Pimolnat Suntian <pimolnats@ecosoft.co.th>
