+++ stock_picking_back2draft/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+This module allows to bring cancelled pickings back to draft
