+++ stock_picking_back2draft/readme/USAGE.rst
@@ -1,0 +1,1 @@
+Just open a cancelled picking and click on 'back to draft' button
