+++ stock_picking_back2draft/views/picking_view.xml
@@ -1,12 +1,18 @@
-<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8" ?>
 <odoo>
     <record id="view_picking_form_reopen" model="ir.ui.view">
         <field name="name">view_picking_form_reopen</field>
         <field name="model">stock.picking</field>
-        <field name="inherit_id" ref="stock.view_picking_form"></field>
+        <field name="inherit_id" ref="stock.view_picking_form" />
         <field name="arch" type="xml">
             <xpath expr="//button[@name='action_cancel']" position="after">
-                <button name="action_back_to_draft" states="cancel" string="Back to draft" groups="base.group_user" type="object"/>
+                <button
+                    name="action_back_to_draft"
+                    states="cancel"
+                    string="Back to draft"
+                    groups="base.group_user"
+                    type="object"
+                />
             </xpath>
         </field>
     </record>
